feat: Add functionality for game logic
I have done several features for this game. I have created different
functions to generate random positions and random values. I have also
create functions to start and clean the game.

The most important changes are the functions related to movement, I abstracted the logic for movement and merging
in order to reuse it for all movement directions.